I'm committing the latest state that I got the code into when working to improve the system's performance.
I'm going to give up on the subparallelograms idea, because the code
becomes too complicated. Unfortunately, I did not have time to
determine whether the subp code would actually be faster. This is an
indication of how bloody complex it was! I think the dvecs approach
seems to work well, and I plan to work this into the master branch of
NeoArealize.